+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed closure bug where multiple tools would return the last tool's content instead of their own content
+
 ## [0.1.1] - 2025-10-20
 
 ### Added